+++ examples/cpp/mnist/train_mnist.cc
@@ -85,7 +85,7 @@ int main(int argc, char** argv) {
   vector<unsigned> cur_labels;
 
   // Run for the given number of epochs (or indefinitely if params.NUM_EPOCHS is negative)
-  while (epoch < params.NUM_EPOCHS || params.NUM_EPOCHS < 0) {
+  while (static_cast<int>(epoch) < params.NUM_EPOCHS || params.NUM_EPOCHS < 0) {
     // Reshuffle the dataset
     cerr << "**SHUFFLE\n";
     random_shuffle(order.begin(), order.end());
